docs: note scraper and poster internal packages
