+++ README.md
@@ -5,6 +5,8 @@
 - `scraper` برای جمع‌آوری خبرهای وب‌سایت CBC Local و ذخیره در MySQL
 - `poster` برای ارسال خبرهای تایید شده به کانال تلگرام
 
+منطق هر کدام در پکیج‌های داخلی `internal/scraper` و `internal/poster` سازمان‌دهی شده است.
+
 ## راه‌اندازی جدول دیتابیس
 
 ```sql
